+++ tests/testthat/test-tar_envvars.R
@@ -3,7 +3,8 @@ tar_test("tar_envvars", {
   expect_equal(sort(colnames(out)), sort(c("name", "value")))
   expect_true(
     all(
-      c("TAR_ASK", "TAR_CONFIG", "TAR_PROJECT", "TAR_WARN") %in% out$name
+      c("TAR_ACTIVE", "TAR_ASK", "TAR_CONFIG", "TAR_PROJECT", "TAR_WARN") %in%
+        out$name
     )
   )
   value <- Sys.getenv("TAR_ASK")